# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Restoran va Kafe CRM/POS tizimi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /venues:
    get:
      operationId: listVenues
      tags: [venues]
      summary: Barcha kafe/restoranlar ro'yxati
      responses:
        "200":
          description: Kafe/restoranlar ro'yxati
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venue"
    post:
      operationId: createVenue
      tags: [venues]
      summary: Yangi kafe/restoran qo'shish (owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenueInput"
      responses:
        "201":
          description: Yaratildi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
  /venues/{id}:
    get:
      operationId: getVenue
      tags: [venues]
      summary: Kafe/restoran ma'lumotlari
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Kafe/restoran
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "404":
          description: Topilmadi
    patch:
      operationId: updateVenue
      tags: [venues]
      summary: Kafe/restoran tahrirlash
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenueUpdate"
      responses:
        "200":
          description: Yangilandi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
    delete:
      operationId: deleteVenue
      tags: [venues]
      summary: Kafe/restoran o'chirish
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: O'chirildi
  /venues/{id}/assign-admin:
    post:
      operationId: assignVenueAdmin
      tags: [venues]
      summary: Kafe uchun admin tayinlash
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAssignInput"
      responses:
        "200":
          description: Admin tayinlandi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /venues/{id}/stats:
    get:
      operationId: getVenueStats
      tags: [venues]
      summary: Kafe statistikasi (bugungi savdo, umumiy qarz)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Statistika
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueStats"
  /venues/{venueId}/products:
    get:
      operationId: listProducts
      tags: [products]
      summary: Kafe mahsulotlari ro'yxati
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Mahsulotlar ro'yxati
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Yangi mahsulot qo'shish
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Mahsulot yaratildi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /venues/{venueId}/products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Mahsulotni tahrirlash
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Yangilandi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Mahsulotni o'chirish
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: O'chirildi
...
  /venues/{venueId}/summary:
    get:
      operationId: getVenueSummary
      tags: [venues]
      summary: Dashboard uchun bugungi savdo xulosasi
      parameters:
        - name: venueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Xulosa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, role]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [owner, admin]
        venueId:
          type: ["integer", "null"]
        venueName:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Venue:
      type: object
      required: [id, name, type]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [cafe, restaurant]
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        adminId:
          type: ["integer", "null"]
        adminName:
          type: ["string", "null"]
        createdAt:
          type: string
    VenueInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [cafe, restaurant]
        address:
          type: string
        phone:
          type: string
    VenueUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [cafe, restaurant]
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
    AdminAssignInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
    VenueStats:
      type: object
      required: [todaySales, totalRevenue, totalDebts, productCount, orderCount]
      properties:
        todaySales:
          type: number
        totalRevenue:
          type: number
        totalDebts:
          type: number
        productCount:
          type: integer
        orderCount:
          type: integer
    Product:
      type: object
      required: [id, venueId, name, price, category]
      properties:
        id:
          type: integer
        venueId:
          type: integer
        name:
          type: string
        price:
          type: number
        category:
          type: string
        description:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, price, category]
      properties:
        name:
          type: string
        price:
          type: number
        category:
          type: string
        description:
          type: string
        isAvailable:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        category:
          type: string
        description:
          type: ["string", "null"]
        isAvailable:
          type: boolean
...
    Order:
      type: object
      required: [id, venueId, totalAmount, paymentType, status, createdAt]
      properties:
        id:
          type: integer
        venueId:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        totalAmount:
          type: number
        paymentType:
          type: string
          enum: [cash, debt]
        status:
          type: string
          enum: [completed, debt]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    DashboardSummary:
      type: object
      required: [todayRevenue, todayOrderCount, totalDebt, unpaidDebtCount, topProducts, recentOrders]
      properties:
        todayRevenue:
          type: number
        todayOrderCount:
          type: integer
        totalDebt:
          type: number
        unpaidDebtCount:
          type: integer
        topProducts:
          type: array
          items:
            $ref: "#/components/schemas/TopProduct"
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
    TopProduct:
      type: object
      required: [productId, productName, totalSold, revenue]
      properties:
        productId:
          type: integer
        productName:
          type: string
        totalSold:
          type: integer
        revenue:
          type: number
...