# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restoran va Kafe CRM/POS tizimi API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  price: number;
  category: string;
  description?: string;
  isAvailable?: boolean;
}