# lib/api-zod/src/generated/types/venue.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restoran va Kafe CRM/POS tizimi API
 * OpenAPI spec version: 0.1.0
 */
import type { VenueType } from './venueType';

export interface Venue {
  id: number;
  name: string;
  type: VenueType;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  adminId?: number | null;
  /** @nullable */
  adminName?: string | null;
  createdAt?: string;
}