# lib/api-zod/src/generated/types/venueInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restoran va Kafe CRM/POS tizimi API
 * OpenAPI spec version: 0.1.0
 */
import type { VenueInputType } from './venueInputType';

export interface VenueInput {
  name: string;
  type: VenueInputType;
  address?: string;
  phone?: string;
}