# lib/api-zod/src/generated/types/venueUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restoran va Kafe CRM/POS tizimi API
 * OpenAPI spec version: 0.1.0
 */
import type { VenueUpdateType } from './venueUpdateType';

export interface VenueUpdate {
  name?: string;
  type?: VenueUpdateType;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
}